requirements.txt
flask==3.1.2
flask-cors
flask-bcrypt
pymongo
python-dotenv
anthropic
gunicorn
dnspython
